app/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
django-import_export
pandas